Use only one file for each model file widget
