+++ moxin-frontend/src/landing/mod.rs
@@ -3,6 +3,8 @@ pub mod downloads;
 pub mod landing_screen;
 pub mod model_card;
 pub mod model_files_list;
+pub mod model_files_items;
+pub mod model_files_tags;
 pub mod model_list;
 pub mod search_bar;
 pub mod search_loading;
