+++ moxin-frontend/src/landing/model_files_list.rs
@@ -1,10 +1,7 @@
-use crate::{
-    data::store::{ModelWithPendingDownloads, Store, StoreAction},
-    shared::{modal::ModalAction, utils::format_model_size},
-};
+use crate::data::store::{ModelWithPendingDownloads, StoreAction};
 use makepad_widgets::*;
-use moxin_protocol::data::{File, FileID, Model, PendingDownload};
-use std::collections::HashMap;
+
+use super::model_files_items::ModelFilesItemsWidgetExt;
 
 live_design! {
     import makepad_widgets::base::*;
@@ -16,8 +13,10 @@ live_design! {
     import crate::shared::widgets::*;
     import crate::landing::shared::*;
 
-    ICON_DOWNLOAD = dep("crate://self/resources/icons/download.svg")
-    ICON_DOWNLOAD_DONE = dep("crate://self/resources/icons/download_done.svg")
+    import crate::landing::model_files_items::ModelFilesItems;
+    import crate::landing::model_files_items::ModelFilesRow;
+    import crate::landing::model_files_tags::ModelFilesTags;
+
     ICON_ADD = dep("crate://self/resources/icons/add.svg")
     ICON_REMOVE = dep("crate://self/resources/icons/remove.svg")
 
@@ -35,174 +34,6 @@ live_design! {
         cell2 = <View> { width: 140, height: 56, padding: 10, align: {x: 0.0, y: 0.5} }
         cell3 = <View> { width: 340, height: 56, padding: 10, align: {x: 0.0, y: 0.5} }
         cell4 = <View> { width: 250, height: 56, padding: 10, align: {x: 0.0, y: 0.5} }
-    }
-
-    ModelFilesListLabel = <RoundedView> {
-        width: Fit,
-        height: Fit,
-        padding: {top: 6, bottom: 6, left: 10, right: 10}
-
-        draw_bg: {
-            instance radius: 2.0,
-            color: #E6F1EC,
-        }
-
-        label = <Label> {
-            draw_text:{
-                text_style: <REGULAR_FONT>{font_size: 9},
-                color: #1C1917
-            }
-        }
-    }
-
-    ModelCardButton = <RoundedView> {
-        width: 140,
-        height: 32,
-        align: {x: 0.5, y: 0.5}
-        spacing: 6,
-
-        draw_bg: { color: #099250 }
-
-        button_icon = <Icon> {
-            draw_icon: {
-                fn get_color(self) -> vec4 {
-                    return #fff;
-                }
-            }
-            icon_walk: {width: Fit, height: Fit}
-        }
-
-        button_label = <Label> {
-            draw_text: {
-                text_style: <REGULAR_FONT>{font_size: 9},
-                fn get_color(self) -> vec4 {
-                    return #fff;
-                }
-            }
-        }
-    }
-
-    DownloadButton = <ModelCardButton> {
-        cursor: Hand,
-        button_label = { text: "Download" }
-        button_icon = { draw_icon: {
-            svg_file: (ICON_DOWNLOAD),
-        }}
-    }
-
-    DownloadedButton = <ModelCardButton> {
-        draw_bg: { color: #fff, border_color: #099250, border_width: 0.5}
-        button_label = {
-            text: "Downloaded"
-            draw_text: {
-                fn get_color(self) -> vec4 {
-                    return #099250;
-                }
-            }
-        }
-        button_icon = {
-            draw_icon: {
-                svg_file: (ICON_DOWNLOAD_DONE),
-                fn get_color(self) -> vec4 {
-                    return #099250;
-                }
-            }
-        }
-    }
-
-    // TODO This is a very temporary solution, we will have a better way to handle this.
-    DownloadPendingButton = <ModelCardButton> {
-        draw_bg: { color: #fff, border_color: #x155EEF, border_width: 0.5}
-        button_label = {
-            text: "Downloading..."
-            draw_text: {
-                fn get_color(self) -> vec4 {
-                    return #x155EEF;
-                }
-            }
-        }
-        button_icon = {
-            draw_icon: {
-                fn get_color(self) -> vec4 {
-                    // invisible for now
-                    return #0000;
-                }
-            }
-        }
-    }
-
-    ModelFilesTags = {{ModelFilesTags}} {
-        width: Fit,
-        height: Fit,
-        flow: Right,
-        spacing: 5,
-
-        template: <ModelFilesListLabel> {}
-    }
-
-    ModelFilesRowWithData = <ModelFilesRow> {
-        show_bg: true,
-        draw_bg: {
-            color: #fff
-        }
-
-        cell1 = {
-            spacing: 10,
-            filename = <Label> {
-                draw_text:{
-                    text_style: <BOLD_FONT>{font_size: 9},
-                    color: #000
-                }
-            }
-        }
-
-        cell2 = {
-            full_size = <Label> {
-                draw_text:{
-                    text_style: <REGULAR_FONT>{font_size: 9},
-                    color: #000
-                }
-            }
-        }
-
-        cell3 = {
-            spacing: 6,
-            quantization_tag = <RoundedView> {
-                width: Fit,
-                height: Fit,
-                padding: {top: 6, bottom: 6, left: 10, right: 10}
-
-                draw_bg: {
-                    instance radius: 2.0,
-                    border_color: #B4B4B4,
-                    border_width: 0.5,
-                    color: #FFF,
-                }
-
-                quantization = <Label> {
-                    draw_text:{
-                        text_style: <REGULAR_FONT>{font_size: 9},
-                        color: #000
-                    }
-                }
-            }
-            tags = <ModelFilesTags> {}
-        }
-
-        cell4 = {
-            align: {x: 0.5, y: 0.5},
-            download_button = <DownloadButton> { visible: false }
-            downloaded_button = <DownloadedButton> { visible: false }
-            download_pending_button = <DownloadPendingButton> { visible: false }
-        }
-    }
-
-    ModelFilesItems = {{ModelFilesItems}} {
-        width: Fill,
-        height: Fit,
-        flow: Down,
-
-        template: <ModelFilesRowWithData> {}
     }
 
     FooterLink = <View> {
@@ -330,261 +161,6 @@ live_design! {
                     apply: {show_all_animation_progress: 0.0}
                 }
             }
-        }
-    }
-}
-
-#[derive(Clone, DefaultNone, Debug)]
-pub enum ModelFileItemsAction {
-    Download(File, Model),
-    Downloaded(FileID),
-    None,
-}
-
-#[derive(Live, LiveHook, LiveRegisterWidget, WidgetRef)]
-pub struct ModelFilesItems {
-    #[rust]
-    area: Area,
-
-    #[walk]
-    walk: Walk,
-
-    #[layout]
-    layout: Layout,
-
-    #[live]
-    template: Option<LivePtr>,
-
-    #[live(true)]
-    show_tags: bool,
-
-    #[live(false)]
-    show_featured: bool,
-
-    #[live(true)]
-    visible: bool,
-
-    #[rust]
-    items: ComponentMap<LiveId, WidgetRef>,
-
-    #[rust]
-    map_to_files: HashMap<LiveId, File>,
-
-    #[rust]
-    model: Option<Model>,
-}
-
-impl Widget for ModelFilesItems {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        let widget_uid = self.widget_uid();
-        let store: &mut Store = scope.data.get_mut::<Store>().unwrap();
-
-        // Notify of a downloaded file.
-        if let Event::Signal = event {
-            if let Some(downloaded_file) = store.downloaded_files_to_notify.pop_front() {
-                    cx.widget_action(
-                        widget_uid,
-                        &scope.path,
-                        ModalAction::ShowModalView(live_id!(popup_download_success_modal_view)),
-                    );
-                    cx.widget_action(
-                        widget_uid,
-                        &scope.path,
-                        ModelFileItemsAction::Downloaded(downloaded_file.clone()),
-                    );
-            }
-        }
-
-        for (id, item) in self.items.iter_mut() {
-            let actions = cx.capture_actions(|cx| item.handle_event(cx, event, scope));
-            if let Some(fd) = item.view(id!(download_button)).finger_down(&actions) {
-                if fd.tap_count == 1 {
-                    let widget_uid = item.widget_uid();
-                    cx.widget_action(
-                        widget_uid,
-                        &scope.path,
-                        ModelFileItemsAction::Download(
-                            self.map_to_files.get(id).unwrap().clone(),
-                            self.model.clone().unwrap(),
-                        ),
-                    );
-
-                    break;
-                }
-            }
-        }
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        let ModelWithPendingDownloads {
-            model,
-            pending_downloads,
-        } = scope.data.get::<ModelWithPendingDownloads>().unwrap();
-        let files = if self.show_featured {
-            Store::model_featured_files(model)
-        } else {
-            Store::model_other_files(model)
-        };
-        cx.begin_turtle(walk, self.layout);
-
-        self.model = Some(model.clone());
-        self.draw_files(cx, &files, pending_downloads);
-        cx.end_turtle_with_area(&mut self.area);
-
-        DrawStep::done()
-    }
-}
-
-impl WidgetNode for ModelFilesItems {
-    fn walk(&mut self, _cx: &mut Cx) -> Walk {
-        self.walk
-    }
-
-    fn redraw(&mut self, cx: &mut Cx) {
-        self.area.redraw(cx)
-    }
-
-    fn find_widgets(&mut self, path: &[LiveId], cached: WidgetCache, results: &mut WidgetSet) {
-        for item in self.items.values_mut() {
-            item.find_widgets(path, cached, results);
-        }
-    }
-}
-
-impl ModelFilesItems {
-    fn draw_files(
-        &mut self,
-        cx: &mut Cx2d,
-        files: &Vec<File>,
-        pending_downloads: &Vec<PendingDownload>,
-    ) {
-        // TODO check if using proper ids in the items collections is better than having this mapping
-        self.map_to_files.clear();
-
-        for i in 0..files.len() {
-            let item_id = LiveId(i as u64).into();
-
-            let item_widget = self
-                .items
-                .get_or_insert(cx, item_id, |cx| WidgetRef::new_from_ptr(cx, self.template));
-            self.map_to_files.insert(item_id, files[i].clone());
-
-            let filename = &files[i].name;
-            let size = format_model_size(&files[i].size).unwrap_or("-".to_string());
-            let quantization = &files[i].quantization;
-            item_widget.apply_over(
-                cx,
-                live! {
-                    cell1 = {
-                        filename = { text: (filename) }
-                    }
-                    cell2 = { full_size = { text: (size) }}
-                    cell3 = {
-                        quantization_tag = { quantization = { text: (quantization) }}
-                    }
-                },
-            );
-
-            if pending_downloads
-                .iter()
-                .find(|f| f.file.id == files[i].id)
-                .is_some()
-            {
-                item_widget.apply_over(
-                    cx,
-                    live! { cell4 = {
-                        download_pending_button = { visible: true }
-                        downloaded_button = { visible: false }
-                        download_button = { visible: false }
-                    }},
-                );
-            } else if files[i].downloaded {
-                item_widget.apply_over(
-                    cx,
-                    live! { cell4 = {
-                        download_pending_button = { visible: false }
-                        downloaded_button = { visible: true }
-                        download_button = { visible: false }
-                    }},
-                );
-            } else {
-                item_widget.apply_over(
-                    cx,
-                    live! { cell4 = {
-                        download_pending_button = { visible: false }
-                        downloaded_button = { visible: false }
-                        download_button = { visible: true }
-                    }},
-                );
-            };
-
-            if self.show_tags {
-                item_widget
-                    .model_files_tags(id!(tags))
-                    .set_tags(cx, &files[i].tags);
-            }
-
-            let _ = item_widget.draw_all(cx, &mut Scope::empty());
-        }
-    }
-}
-
-impl ModelFilesItemsRef {
-    fn get_height(&mut self, cx: &mut Cx) -> f64 {
-        let Some(inner) = self.borrow_mut() else {
-            return 0.0;
-        };
-        inner.area.rect(cx).size.y
-    }
-}
-
-#[derive(Live, LiveHook, Widget)]
-pub struct ModelFilesTags {
-    #[redraw]
-    #[rust]
-    area: Area,
-
-    #[walk]
-    walk: Walk,
-
-    #[layout]
-    layout: Layout,
-
-    #[live]
-    template: Option<LivePtr>,
-
-    #[rust]
-    items: ComponentMap<LiveId, WidgetRef>,
-}
-
-impl Widget for ModelFilesTags {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        for (_id, item) in self.items.iter_mut() {
-            item.handle_event(cx, event, scope);
-        }
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        cx.begin_turtle(walk, self.layout);
-        for (_id, item) in self.items.iter_mut() {
-            let _ = item.draw_walk(cx, scope, walk);
-        }
-        cx.end_turtle_with_area(&mut self.area);
-        DrawStep::done()
-    }
-}
-
-impl ModelFilesTagsRef {
-    pub fn set_tags(&self, cx: &mut Cx, tags: &Vec<String>) {
-        let Some(mut tags_widget) = self.borrow_mut() else {
-            return;
-        };
-        tags_widget.items.clear();
-        for (i, tag) in tags.iter().enumerate() {
-            let item_id = LiveId(i as u64).into();
-            let item_widget = WidgetRef::new_from_ptr(cx, tags_widget.template);
-            item_widget.apply_over(cx, live! {label = { text: (tag) }});
-            tags_widget.items.insert(item_id, item_widget);
         }
     }
 }
